Update package.json and Astro components
This commit updates the "dev" script in the package.json file to include the "astro check" command before running "astro dev". It also updates the Props interface in the ListItem.astro component to make the "title" and "description" properties optional. In the Post.astro component, it updates the import statement for the "formatDate" function and the "translate" variable. It also adds line breaks for better readability. In the config.ts file, it renames the "postsCollection" variable to "posts". In the archive.astro file, it updates the logic for getting the year from the post's pubDate. In the atom.xml.ts file, it adds more properties to the generated feed items. Finally, in the utils/index.ts file, it updates the formatDate function to handle cases where the date is undefined.